add user registration service changes

+++ BlogSiteWebAPI/BlogInfoService/Services/BlogDetailsService.cs
@@ -9,21 +9,40 @@ namespace BlogInfoService.Services
         public BlogDetailsService(IBlogStoreDBSetting settings, IMongoClient mongoClient)
         {
             var database = mongoClient.GetDatabase(settings.DatabaseName);
-            _blogDetails = database.GetCollection<BlogDetails>(settings.UserRegCollectionName);
+            _blogDetails = database.GetCollection<BlogDetails>(settings.BlogDtlsCollectionName);
         }
         public int AddBlog(BlogDetails blogDtls)
         {
-            _blogDetails.InsertOne(blogDtls);
-            return 1;
+            try
+            {
+                _blogDetails.InsertOne(blogDtls);
+                return 1;
+            }
+            catch (Exception)
+            {
+                return 0;
+            }
+
         }
-        public void DeleteBlog(int blogId)
+        public int DeleteBlog(int blogId)
         {
-            _blogDetails.DeleteOne(blog => blog.BlogId == blogId);
+            try
+            {
+                var itemToRemove = _blogDetails.Find(blog => blog.BlogId == blogId).SingleOrDefault();
+                if (itemToRemove != null)
+                {
+                    _blogDetails.DeleteOne(blog => blog.BlogId == blogId);
+                    return 1;
+                }
+                else
+                {
+                    return 0;
+                }
+            }
+            catch (Exception)
+            {
+                return 0;
+            }
         }
-        
-        //public BlogDetails GetBlogById(int blogId)
-        //{
-        //    return _blogDetails.Find(blog => blog.BlogId==blogId).FirstOrDefault();
-        //}
     }
 }
